requirements.txt
gradio
insightface
opencv-python-headless
onnxruntime
scikit-learn
joblib
Pillow
